requirements.txt
python-esl
redis
requests
flask
flask-restful
